fixed moneyReturn methods

+++ Supermarket/src/classes/OrdinaryClient.java
@@ -49,7 +49,7 @@ public class OrdinaryClient extends Actor {
     }
 
     @Override
-    public float moneyReturn(int orderPrice) {
+    public float moneyReturn(float orderPrice) {
         return (float) (orderPrice * 0.5);
     }
 }
